accounts-db: Moves test helpers out of impl.rs
